Revert "refactor(color): carve RhiToneMapper into packages/color/, engine stays color-pure"
This reverts commit 673d0da70948dc55ebbf3cf5e2d19ceecfbfaf3e.
