+++ libs/streamlib-engine/src/core/color/mod.rs
@@ -13,11 +13,13 @@
 mod matrix;
 mod resolve;
 mod resolved;
+mod tone;
 mod transfer;
 
 pub use matrix::{yuv_to_rgb_matrix, YuvToRgbDecomposition};
 pub use resolve::resolve_color_defaults;
 pub use resolved::{ColorSpaceKind, ResolvedColorInfo};
+pub use tone::{bt2390_eetf_per_channel, bt2446a_inverse_per_channel};
 pub use transfer::{
     bt709_to_linear, from_linear, hlg_to_linear, linear_to_bt709, linear_to_hlg, linear_to_pq,
     linear_to_srgb, pq_to_linear, srgb_to_linear, to_linear, TransferId,
