+++ libs/streamlib-engine/src/vulkan/rhi/mod.rs
@@ -48,6 +48,11 @@ pub use streamlib_consumer_rhi::{
 mod vulkan_blitter;
 pub use vulkan_blitter::VulkanBlitter;
 
+#[cfg(target_os = "linux")]
+mod vulkan_tone_mapper;
+#[cfg(target_os = "linux")]
+pub use vulkan_tone_mapper::{VulkanToneMapper, TONE_MAPPER_WORKGROUP_SIZE};
+
 pub(crate) mod vulkan_buffer;
 pub use vulkan_buffer::HostVulkanBuffer;
